Audit fixes + features: intervals, session totals, 64-bit counters
Fixes:
- CI now triggers on main (was master); fix RELEASE_NOTES filename case
- Zero the readout when offline instead of showing a stale value
- Drop per-second log spam; rename logger category
- C++ core uses NET_RT_IFLIST2 / if_data64 (64-bit counters, no wrap math)
- Bundle id -> com.araidz.NetSpeedMonitor
- Strip dead OC fields + unused reset; cache SCDynamicStore handle
- Remove stray icon.png; gitignore *.dmg

Features:
- Selectable refresh interval (0.5/1/2/5s), persisted
- Session totals (received/sent) with reset in the menu
- Version line in menu; VoiceOver label on the icon; idle opacity floor

Bump to 1.1. Docs updated.

+++ NetSpeedMonitor/NetTrafficStatCpp.hpp
@@ -10,8 +10,8 @@ using time_point_type = std::chrono::time_point<clock_type, duration_type>;
 
 struct NetTrafficStat {
     time_point_type tp_retrieval; // time_point where this stat is retrieved
-    uint32_t ifi_ibytes = 0;      // raw ifi_ibytes message
-    uint32_t ifi_obytes = 0;      // raw ifi_obytes message
+    uint64_t ifi_ibytes = 0;      // raw ifi_ibytes message (64-bit, from if_data64)
+    uint64_t ifi_obytes = 0;      // raw ifi_obytes message (64-bit, from if_data64)
     int64_t total_ibytes = 0;     // i.e. ifi_ibytes accumulated in wider int type
     int64_t total_obytes = 0;     // i.e. ifi_obytes accumulated in wider int type
